unify image, audio, video behavior
